fixed int error in decimate_matrix

+++ miprest.py
@@ -22,7 +22,7 @@ def decimate_matrix(X,dfactor):
     columns removed.  So if dfactor = 4 and X is 10 x 1000, the matrix
     returned will be size 10 x 250.
     '''
-    ndrop = X.shape[1] - ceil(X.shape[1]/dfactor)
+    ndrop = int(X.shape[1] - ceil(X.shape[1]/dfactor))
     dropcol = permutation(X.shape[1])[:ndrop]
     return delete(X,dropcol,1)
